project1/requirements.txt
aws-cdk.core
aws-cdk.aws-s3
aws-cdk.aws-ssm
aws-cdk.aws-ec2
# aws-cdk.aws-kms
# aws-cdk.aws-iam
# aws-cdk.aws-secretsmanager
# aws-cdk.aws-elasticsearch
# aws-cdk.aws-cognito
# aws-cdk.aws-apigateway
# aws-cdk.aws-lambda
# aws-cdk.aws-sns
# aws-cdk.aws-events
# aws-cdk.aws-events-targets
# aws-cdk.aws-wafv2
# aws-cdk.aws-route53
# aws-cdk.aws-route53-targets
# aws-cdk.aws-cloudtrail